Add new structure to support bazel credentials.
