+++ cmd/docker/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/docker/docker-credential-helpers/credentials"
+
+func main() {
+	credentials.Serve(ArtifactoryKeychain{})
+}
