+++ cmd/docker/main_test.go
@@ -1,0 +1,14 @@
+package main
+
+import (
+	"os"
+	"testing"
+)
+
+func TestCommandList(t *testing.T) {
+	os.Args = []string{
+		"docker-credential-helper",
+		"list",
+	}
+	main()
+}
